Minor Edits to Ivelin's PCA
I added a significance test to the CV. I found that the minimized root PRESS statistic is not statistically different for 6 principal components from 7 or 8 so it automatically pared down the model to 6 components. This improved the Kaggle score a little bit so I made my interpretation using this change.